+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/settings/ProfilesSettingsViewModel.kt
@@ -100,7 +100,7 @@ class ProfilesSettingsViewModelImpl(
 
     override fun closeProfile() {
         log.debug { "close profile" }
-        coroutineScope.launch { profileManager.closeProfile() }
+        profileManager.closeProfile()
     }
 
     override fun setMultiProfileEnabled(enabled: Boolean) {